requirements.txt
azure-identity
azure-ai-projects
azure-ai-inference
azure-ai-evaluation
azure-ai-contentsafety
azure-monitor-opentelemetry
azure-ai-inference[opentelemetry]
jupyter
python-dotenv
notebook
ipykernel
pandas
tabulate
nbformat
asyncio
typing
mkdocs-material
mkdocs-render-swagger-plugin
mkdocs-macros-plugin
mkdocs-with-pdf
mkdocs-glightbox
mkdocs-open-in-new-tab
mkdocs-video
mkdocs-table-reader-plugin
mkdocs-minify-plugin
mkdocs-jupyter
semantic-kernel
mkdocs-git-committers-plugin-2
mkdocs-git-revision-date-localized-plugin
mkdocs-git-revision-date-plugin
